console/requirements.txt
paho-mqtt
tabulate